Moved tests around.
