+++ test/parameters/NumberParameters.spec.ts
@@ -1,7 +1,7 @@
 import { expect } from "chai";
 import "mocha";
-import { createRegex, test } from "./_parameter-testing";
-import { NumberParameter, NumberStyle, FractionParameter } from "../src";
+import { createRegex, test } from "../_parameter-testing";
+import { NumberParameter, NumberStyle, FractionParameter } from "../../src";
 
 describe("NumberParameters.param.spec.ts / Default commands", () => {
