+++ test/parameters/StringParameters.spec.ts
@@ -3,12 +3,12 @@ import {
   ChoiceParameter,
   TokenParameter,
   IPv4Parameter
-} from "./../src/";
+} from "../../src";
 
-import { RestParameter } from "./../src/index";
+import { RestParameter } from "../../src/index";
 import { expect } from "chai";
 import "mocha";
-import { createRegex, test } from "./_parameter-testing";
+import { createRegex, test } from "../_parameter-testing";
 
 describe("StringParameters.spec.ts", () => {
   describe("RestParameter", () => {
