+++ test/same-start-name.spec.ts
@@ -14,9 +14,9 @@ describe("same-start-name.spec.ts > execute commands with the same start name", 
     var options = new Options();
     options.verbose = false;
 
-    map_command(pretend.robot, "ci", (context: IContext) => context.res.reply("ci"), options);
-    map_command(pretend.robot, "cd", (context: IContext) => context.res.reply("cd"), options);
-    map_command(pretend.robot, "cicd", (context: IContext) => context.res.reply("cicd"), options);
+    map_command(pretend.robot, "ci", (context) => context.res.reply("ci"), options);
+    map_command(pretend.robot, "cd", (context) => context.res.reply("cd"), options);
+    map_command(pretend.robot, "cicd", (context) => context.res.reply("cicd"), options);
 
   });
